lms-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
fastapi-users[sqlalchemy]==12.1.3
alembic==1.13.1
# psycopg2 is already available on system
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
# Use pydantic versions that work
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
# asyncpg==0.29.0  # commented out until we can install without compilation 